Moved tools into lib, removed unused tools
