+++ source/LoomSE/lib/common/index.js
