+++ source/LoomSE/lib/index.js
@@ -1,3 +1,4 @@
+export * from './browser';
 export * from './common';
 export * from './fileParsers';
 export * from './time';
