+++ source/LoomSE/lib/time/index.js
