--- source/LoomSE/tools/common/index.js
@@ -1,1 +1,0 @@
-export * from './common';
