--- source/LoomSE/tools/fileParsers/index.js
@@ -1,1 +1,0 @@
-export * from './fileParsers';
