--- source/LoomSE/tools/fileParsers/parsers/index.js
@@ -1,1 +1,0 @@
-export { SRT_Parser } from './SRT_Parser';
