--- source/LoomSE/tools/time/index.js
@@ -1,1 +1,0 @@
-export * from './time';
